Refactor email template functions to the TemplateProcessor
